refactor (wip)
